Rename package webservlet to webannotation
