+++ extension/tomcat9/src/main/java/cloud/piranha/extension/tomcat9/Tomcat9Extension.java
@@ -33,7 +33,7 @@ import cloud.piranha.pages.jasper.JasperExtension;
 import cloud.piranha.webapp.annotationscan.AnnotationScanExtension;
 import cloud.piranha.webapp.scinitializer.ServletContainerInitializerExtension;
 import cloud.piranha.webapp.tempdir.TempDirExtension;
-import cloud.piranha.webapp.webservlet.WebAnnotationExtension;
+import cloud.piranha.webapp.webannotation.WebAnnotationExtension;
 import cloud.piranha.webapp.webxml.WebXmlExtension;
 
 /**
