+++ extension/webprofile/src/main/java/cloud/piranha/extension/webprofile/WebProfileExtension.java
@@ -32,7 +32,7 @@ import cloud.piranha.webapp.api.WebApplicationExtensionContext;
 import cloud.piranha.webapp.annotationscan.AnnotationScanExtension;
 import cloud.piranha.webapp.scinitializer.ServletContainerInitializerExtension;
 import cloud.piranha.webapp.tempdir.TempDirExtension;
-import cloud.piranha.webapp.webservlet.WebAnnotationExtension;
+import cloud.piranha.webapp.webannotation.WebAnnotationExtension;
 import cloud.piranha.webapp.webxml.WebXmlExtension;
 
 /**
